fix!: Preserve null values in `pct_change`
